requirements.txt
opencv-python-headless
Pillow
pillow-heif
numpy
tqdm